qtscriptgenerator: Version bump of sorts;
This is the lastest git pull. It appears from looking that the build .so files are designated version 1.0.0.
So that's what I will call it. Some folks think this is sort of a dead project. I don't know but to be sure
I have put this in the warm bossom of doppio.

And it still needs a patch for qt-4.8.0;
http://code.google.com/p/qtscriptgenerator/issues/detail?id=80#c6 <--- qtscripgenerator patch (see comment #6)