Bumped to 4.3.15.
If you're going to suddenly decide that the "zsh" package needs
to go with the stable branch (after going with the development branch
for so long), it'd be nice to at least maintain the development branch's
version of the module.  This is the first attention that zsh-dev has
gotten since 2009.