Python: removing this "compileall" step. The reason for its original insertion
is lost in /dev/null.

From that script;

"When called as a script with arguments, this compiles the directories
given as arguments recursively; the -l option prevents it from
recursing into directories.

Without arguments, if compiles all modules on sys.path, without
recursing into subdirectories.  (Even though it should do so for
packages -- for now, you'll have to deal with packages separately.)"

I don't see this doing much different than a default_build. Additionally,
on a clean install the make will tank on directory missing. It was on
IRC someone had that problem. I have confirmed this does happen in such a
situation.