polkit: Version bump. There have been changes in file locations with versions 0.109 and
greater, so adjusting the BUILD to reflect that (see the comments within the BUILD). Notabley
it does not use /var. Additionally, versions 0.109 require a polkitd user. Just to make sure
the old dirs/files are removed I have used the hammer.

Editing the polkit.service and getting rid of the --no-debug as it wouldn't even tell you
it wants/needs a "home". Ratler noted other distros use / and so shall we unless there
are objections.

Based on input from Ratler our pam guru, inserting an additional pam rule to its
/etc/pam.d/polkit-1.