poppler: fix a problem with the cms option
to turn on/off the cms the syntax is (now) "--enable-cms=[auto|lcms1|lcms2|none]"
And the syntax in the DEPENDS file is "--enable-cms / --disable-cms" which both are treated as "--enable-cms=none" by the configure script.
This syntax breaks modules that need poppler to be compiled with cmsi (like cups-filters).