httpd: roll back MAJOR version change to 2.4, bump to 2.2.24
Bumping to version 2.4 of httpd is a HUGE change and breaks MANY
people's servers.  Don't do that.  This is the second time I've had to
do this exact same commit.  Stop just bumping to whatever the biggest
number is by reflex.  Apache 2.4 is different enough that it practically
justifies a new module.

Stop breaking people's servers.  For crying out loud.