freetype2: version bumped to 2.5.2
Re-worked the module slightly. Some of the seds are now part of
two of the patches. A couple of old seds was invalid and
TT_CONFIG_OPTION_UNPATENTED_HINTING was not a recommended setting unless
you use a few specific asian fonts, for the rest it will just cause issues
instead.