slim: make the first item in sessions always an empty default
This solution is specific to lunar and authored by me.
It is much less intrusive than all the patches I looked at.

However, if a user doesn't handle an empty session parameter
in his ~/.xinitrc script and starts a default in this case,
this (first) entry will break down and he will have to select a
different one.