seamonkey: Bump to 2.26.1.
Moved somethings to PRE_BUILD from BUILD. Needed a CFLAGS
to compile with tree freetype.

A sedit needed for it to find system pixman.h.

Has the same issue as firefox-31.0 requiring --disable-shared-js.

Not catastrophic but removed the switches from CONFIGURE, was appending two of those to mozconfig.

Probably a few other observations I am forgetting but it compiles with yes to all the optional_depends (sans gnome-vfs).

Since sqlite is used to track history and bookmarks, created some history and bookmarks and noted
no issues. Granted it was not an extensive attempt but one none the less, ymmv.