dash: A quick, basic, no-frills shell.
What with a whole collection of security holes in bash recently, I
think that a small, simple, secure shell might be a better
candidate for /bin/sh.  So I'm offering the /bin/sh that some
other distros use to try out for a while.

It installs as /usr/bin/dash for the moment, making it /bin/sh is
something only recommended for adventurous people who don't mind
potentially breaking their system.