Revert "boost: Adding a sed to the BUILD. Taken from http://www.linuxfromscratch.org/blfs/view/svn/general/boost.html"
This reverts commit 714627c69dc4d592d460d94a558243139f4450cf.

I disagree with this commit.
It's not boost's mistake that modules fail to "correctly" use qt4.
The correct fix for the build failures you point out are to add
the respective Q_MOC_RUN safeguards around the boost includes
in the source files of the dependent projects which are passed to qt's moc.