mutt-devel: bump to 1.5.24.
This came out August 31.

The last "stable" release of mutt came out in June 2007.  I wonder what
it'll take to get them to admit that it's not so much "stable" as
"fossilized".

I took out the sidebar patch because I've never used whatever
functionality it provides.  Anyone who actually does use that is welcome
to reimplement it.