opencv-3: There were still a few lib issues between this and opencv. Both insisted on using a
few of the same so names and the mangling did not resolve that aspect, it only dealt with versioning
and linking. So the solution is remove opencv. This requires adjustments to BUILD of those modules
dependent or optionally which I have corrected and they are all fine with opencv-3.

For future information, if you run into an issue it will most likely be related to -lippicv not
found even though it is listed by pkgconfig so you need to fiddle with LDFLAGS.