spidermonkey: Fixed the issue of the missing version numbers.
This is a bit of a terrible hack: the Mozilla developers are assuming an
older version of Perl where defined(@ARRAYNAME) was the way to check
whether an array had anything in it, where in modern Perl, the correct
way to check for that is to evaluate the array in a scalar context.

I tried to make the sedit in a way that it would catch that specific
usage, so that if Mozilla fixes it in the future, the sedit simply won't
change anything any more.

Note that now is installs the JavaScript REPL as "/usr/bin/js24" instead
of "/usr/bin/js".  This might cause problems.