virtuoso: bump to 7.2.4.2. If you say yes to perl and run into this;
/usr/lib/perl5/5.26.1/x86_64-linux-thread-multi/CORE/perl.h:738:13: fatal error: xlocale.h: No such file or directory

That means you went for the glibc-2.26 bump. This is the second module I ran into (well perl) that uses xlocale.h in past
versions of glibc. Well the glibc folks dropped that header and now you only have locale.h. Why, dunno but the short of it,
relin perl and you should be good to go.