httpd: Fix the systemd service file.
You can either make the service type "simple" and start httpd
with the "-DFOREGROUND" flag, or call it "forking", and start
it without that flag.

You can't have it sort of half way.  That just results in an httpd
service that mysteriously times out because it isn't behaving the way
systemd expects it to behave..