VirtualBox: CI failed on; kmk: lrelease: Command not found
which is provided by qt5. DEPENDS will use --disable-qt, even so the
build still looks for lrelease. Qt5 is only for a GUI. As such lets
try chaning that switch to --build-headless and see if CI like that.

Also added a conditional to BUILD to source qt5.rc if using the GUI.