foomatic-db-engine: Bump to 4.0-20231024. We don't use any of the
"current" foomatics because they change often. This would cause
un-needed effort maintaining them, unless you disable sha256, which
we prefer *not* to do.

If you want them bumped, a copy needs to be uploaded to mirrors on
robusta. For those without acces, I'm generally around so message
if in need when you decide to bump any of the foomatic stuff.