rustc: install the correct stage binaries
stage1 binaries are just ONLY to build the second stage of the rustc
compiler. Installing the stage1 files results in broken dependencies.
Stage 2 are the files that should have been installed, and it fixes the
issues with links against wrong and non-existing binaries.

Please use lunar nofix <module> when testing stuff.