XML-SAX-Expat: Add libwww-perl depends. AFAICT this is actually needed
by XML-Parser. Gentoo and Archlinux agree on this. However, XML-Parser
lives in core and I know we try to keep that as small as possible and
libwww-perl drags in a fair number of other perl modules not in core.

If this works then skirts save the day. If not, then I'm outa of
ideas and someone else should look the problem.