sqlite: Remove mandatory dependencies
Being forced to install all of tcl just for one optional component of
sqlite is simply insane. Make it optional so that if you really really
want to run all the tests and/or run sqlite3_analyze, you can, but if
you just need sqlite, you can have it by itself.

sqlite is meant to be small and standalone, not bring in entire other
ecosystems.